データ形式を変換するメソッドを、FileOperationクラスに移動
