+++ 04.class/all_memos.js
@@ -2,18 +2,12 @@ import OneMemo from "./one_memo.js";
 
 export default class AllMemos {
   #memos;
+  constructor(memoObjects) {
+    this.#memos = memoObjects;
+  }
 
   getMemos() {
     return this.#memos;
-  }
-
-  async toObjects(memoJson) {
-    const allLines = JSON.parse(memoJson).allLines;
-    const allMemos = [];
-    await allLines.forEach((lines) => {
-      allMemos.push(new OneMemo(lines));
-    });
-    this.#memos = allMemos;
   }
 
   delete(index) {
@@ -25,16 +19,5 @@ export default class AllMemos {
     const allMemos = this.#memos;
     allMemos.push(newMemo);
     this.#memos = allMemos;
-  }
-
-  async getJson() {
-    const allMemos = this.#memos;
-    const allLines = [];
-    await allMemos.forEach((oneMemo) => {
-      allLines.push(oneMemo.getLines());
-    });
-    const memoFile = { allLines: allLines };
-    const memoJson = JSON.stringify(memoFile, null, 2);
-    return memoJson;
   }
 }
